添加数据库虚拟字段

+++ controller/menu.js
@@ -1,15 +1,13 @@
 const MenuModel = require("../model/menu");
-
+const FoodsModel = require("../model/foods");
 
 module.exports = {
   async queryList(ctx) {
-   
+    const result = (await MenuModel.find().populate("foodsCount").sort({ createdAt: -1 })).map((item) => {
+      return item.toObject({ virtuals: true });
+    });
 
-    const result = await MenuModel.find()
-      .sort({ createdAt: -1 })
-     
-
-    ctx.body = result
+    ctx.body = result;
   },
   async createOne(ctx) {
     const payload = ctx.request.body;
@@ -34,6 +32,16 @@ module.exports = {
   },
   async deleteOne(ctx) {
     const { id } = ctx.params;
+
+    try {
+      var result = await FoodsModel.findOne({ menuID: id });
+    } catch (error) {}
+     
+    if (result) {
+      ctx.status = 403;
+      ctx.body = { message: "此分类下有商品" };
+      return;
+    }
     ctx.body = await MenuModel.findByIdAndRemove(id);
   },
 };
